requirements.txt
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
pytest>=7.0.0
black>=23.0.0
isort>=5.0.0
mypy>=1.0.0
